Moved MetricsServer to metrics module
