+++ inorbit_connector/metrics/__init__.py
@@ -14,7 +14,7 @@ Framework instruments (always declared, no-op when metrics are disabled):
 * ``inorbit.connector.execution_loop.ticks`` — successful run-loop iterations
 * ``inorbit.connector.execution_loop.errors`` — exceptions caught in the loop
 
-For domain metrics (e.g. ``mir.api.errors``, ``mqtt.broker.connected``),
+For domain metrics (e.g. ``fleet.api.errors``, ``mqtt.broker.connected``),
 concrete connectors get their own meter via
 ``inorbit_edge.metrics.get_meter("inorbit_<vendor>_connector")`` and declare
 instruments on it. They share the global MeterProvider installed here, so
@@ -27,7 +27,11 @@ is ``False`` (the default); when False, no provider is installed and no
 HTTP server is started.
 """
 
+import json
 import logging
+import os
+import socket
+from pathlib import Path
 
 from inorbit_edge.metrics import (
     OTEL_API_AVAILABLE,
@@ -35,12 +39,16 @@ from inorbit_edge.metrics import (
     get_meter,
     setup_prometheus_meter_provider,
 )
+from prometheus_client import start_http_server
 
 from inorbit_connector import __version__ as _connector_version
 from inorbit_connector.models import MetricsConfig
 
 
 _logger = logging.getLogger(__name__)
+
+
+# --- Provider setup -------------------------------------------------------
 
 
 def setup_prometheus_metrics(
@@ -75,6 +83,95 @@ def setup_prometheus_metrics(
             "pick up inorbit-edge[telemetry]."
         )
     return installed
+
+
+# --- HTTP server + file_sd discovery file ---------------------------------
+
+
+class MetricsServer:
+    """Prometheus HTTP server + ``file_sd``-format discovery writer.
+
+    The HTTP-serving piece is one call to ``prometheus_client.start_http_server``;
+    the value-add of this class is the discovery file. On ``start()`` the
+    connector writes ``<discovery_dir>/<connector_id>.json`` (atomic
+    tmp-and-rename) describing its bound ``host:port`` so a host-side OTEL
+    collector can pick it up via ``file_sd_configs``. ``stop()`` removes the
+    file and shuts the HTTP server down.
+
+    Both methods degrade silently with a log entry if something goes wrong.
+    """
+
+    def __init__(self, config: MetricsConfig, connector_id: str) -> None:
+        self._config = config
+        self._connector_id = connector_id
+        self._http_server = None
+        self._http_thread = None
+        self.actual_port: int | None = None
+
+    def start(self) -> None:
+        try:
+            self._http_server, self._http_thread = start_http_server(
+                port=self._config.bind_port,
+                addr=self._config.bind_host,
+            )
+            self.actual_port = self._http_server.server_address[1]
+        except Exception as exc:
+            _logger.error(
+                "Failed to bind metrics HTTP server on %s:%s: %s",
+                self._config.bind_host,
+                self._config.bind_port,
+                exc,
+            )
+            self._http_server = None
+            self._http_thread = None
+            self.actual_port = None
+            return
+
+        try:
+            self._write_discovery_file()
+        except Exception as exc:
+            _logger.error("Failed to write metrics discovery file: %s", exc)
+
+    def stop(self) -> None:
+        if self._http_server is not None:
+            try:
+                self._http_server.shutdown()
+            except Exception as exc:
+                _logger.error("Error shutting down metrics HTTP server: %s", exc)
+            self._http_server = None
+            self._http_thread = None
+
+        path = self._discovery_file_path()
+        try:
+            if path.exists():
+                path.unlink()
+        except Exception as exc:
+            _logger.error(
+                "Failed to remove metrics discovery file %s: %s", path, exc
+            )
+
+    def _discovery_file_path(self) -> Path:
+        return Path(self._config.discovery_dir) / f"{self._connector_id}.json"
+
+    def _write_discovery_file(self) -> None:
+        discovery_dir = Path(self._config.discovery_dir)
+        discovery_dir.mkdir(parents=True, exist_ok=True)
+
+        advertise_host = (
+            self._config.advertise_host
+            or socket.gethostname()
+            or "localhost"
+        )
+        body = [
+            {
+                "targets": [f"{advertise_host}:{self.actual_port}"],
+                "labels": {},
+            }
+        ]
+        path = self._discovery_file_path()
+        tmp = path.with_suffix(path.suffix + ".tmp")
+        tmp.write_text(json.dumps(body) + "\n")
+        os.replace(tmp, path)
 
 
 # --- Framework instruments ------------------------------------------------
